dom/presentation/tests/mochitest/test_presentation_1ua_sender_and_receiver_oop.html

<!DOCTYPE HTML>
<!-- vim: set shiftwidth=2 tabstop=2 autoindent cindent expandtab: -->
<html>
<!-- Any copyright is dedicated to the Public Domain.
   - http://creativecommons.org/publicdomain/zero/1.0/ -->
  <head>
    <meta charset="utf-8">
    <title>Test for B2G Presentation API when sender and receiver at the same side (OOP ver.)</title>
    <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
    <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  </head>
  <body>
    <a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1234492">
    Test for B2G Presentation API when sender and receiver at the same side (OOP ver.)</a>
<script type="application/javascript;version=1.8">

'use strict';

var gScript = SpecialPowers.loadChromeScript(SimpleTest.getTestFileURL('PresentationSessionChromeScript1UA.js'));
var receiverUrl = SimpleTest.getTestFileURL('file_presentation_1ua_receiver.html');
var request;
var connection;
var receiverIframe;

function debug(str) {
  // info(str);
}

function postMessageToIframe(aType) {
  receiverIframe.src = receiverUrl + "#" +
                       encodeURIComponent(JSON.stringify({ type: aType }));
}

function setup() {

  gScript.addMessageListener('device-prompt', function devicePromptHandler() {
    debug('Got message: device-prompt');
    gScript.removeMessageListener('device-prompt', devicePromptHandler);
    gScript.sendAsyncMessage('trigger-device-prompt-select');
  });

  gScript.addMessageListener('control-channel-established', function controlChannelEstablishedHandler() {
    debug('Got message: control-channel-established');
    gScript.removeMessageListener('control-channel-established',
                                  controlChannelEstablishedHandler);
    receiverIframe = document.createElement('iframe');
    receiverIframe.setAttribute("remote", "true");
    receiverIframe.setAttribute("mozbrowser", "true");
    receiverIframe.setAttribute("mozpresentation", receiverUrl);
    receiverIframe.setAttribute('src', receiverUrl);
    receiverIframe.addEventListener("mozbrowserloadend", function mozbrowserloadendHander() {
      receiverIframe.removeEventListener("mozbrowserloadend", mozbrowserloadendHander);
      info("Receiver loaded.");
      gScript.sendAsyncMessage("trigger-control-channel-open");
    });

    // This event is triggered when the iframe calls "alert".
    receiverIframe.addEventListener("mozbrowsershowmodalprompt", function receiverListener(evt) {
      var message = evt.detail.message;
      if (/^OK /.exec(message)) {
        ok(true, message.replace(/^OK /, ""));
      } else if (/^KO /.exec(message)) {
        ok(false, message.replace(/^KO /, ""));
      } else if (/^INFO /.exec(message)) {
        info(message.replace(/^INFO /, ""));
      } else if (/^COMMAND /.exec(message)) {
        var command = JSON.parse(message.replace(/^COMMAND /, ""));
        gScript.sendAsyncMessage(command.name, command.data);
      } else if (/^DONE$/.exec(message)) {
        receiverIframe.removeEventListener("mozbrowsershowmodalprompt",
                                            receiverListener);
        teardown();
      }
    }, false);

    var promise = new Promise(function(aResolve, aReject) {
      document.body.appendChild(receiverIframe);
      aResolve(receiverIframe);
    });

    var obs = SpecialPowers.Cc["@mozilla.org/observer-service;1"]
                           .getService(SpecialPowers.Ci.nsIObserverService);
    obs.notifyObservers(promise, 'setup-request-promise', null);
  });

  gScript.addMessageListener('promise-setup-ready', function promiseSetupReadyHandler() {
    debug('Got message: promise-setup-ready');
    gScript.removeMessageListener('promise-setup-ready',
                                  promiseSetupReadyHandler);
    gScript.sendAsyncMessage('trigger-on-session-request', receiverUrl);
  });

  gScript.addMessageListener('offer-sent', function offerSentHandler() {
    debug('Got message: offer-sent');
    gScript.removeMessageListener('offer-sent', offerSentHandler);
    gScript.sendAsyncMessage('trigger-on-offer');
  });

  gScript.addMessageListener('answer-sent', function answerSentHandler() {
    debug('Got message: answer-sent');
    gScript.removeMessageListener('answer-sent', answerSentHandler);
    gScript.sendAsyncMessage('trigger-on-answer');
  });

  return Promise.resolve();
}

function testCreateRequest() {
  return new Promise(function(aResolve, aReject) {
    info('Sender: --- testCreateRequest ---');
    request = new PresentationRequest("http://example.com");
    request.getAvailability()
      .then((aAvailability) => {
        aAvailability.onchange = function() {
          aAvailability.onchange = null;
          ok(aAvailability.value, "Sender: Device should be available.");
          aResolve();
        }
      })
      .catch((aError) => {
        ok(false, "Sender: Error occurred when getting availability: " + aError);
        teardown();
        aReject();
      });

    gScript.sendAsyncMessage('trigger-device-add');
  });
}

function testStartConnection() {
  return new Promise(function(aResolve, aReject) {
    request.start()
    .then((aConnection) => {
      connection = aConnection;
      ok(connection, "Sender: Connection should be available.");
      ok(connection.id, "Sender: Connection ID should be set.");
      is(connection.state, "connected", "Sender: Connection state at sender side should be connected by default.");
      aResolve();
    })
    .catch((aError) => {
      ok(false, "Sender: Error occurred when establishing a connection: " + aError);
      teardown();
      aReject();
    });
  });
}

function testSendMessage() {
  return new Promise(function(aResolve, aReject) {
    info('Sender: --- testSendMessage ---');
    gScript.addMessageListener('trigger-message-from-sender', function triggerMessageFromSenderHandler() {
      gScript.removeMessageListener('trigger-message-from-sender', triggerMessageFromSenderHandler);
      info('Send message to receiver');
      connection.send('msg-sender-to-receiver');
    });

    gScript.addMessageListener('message-from-sender-received', function messageFromSenderReceivedHandler() {
      gScript.removeMessageListener('message-from-sender-received', messageFromSenderReceivedHandler);
      aResolve();
    });
  });
}

function testIncomingMessage() {
  return new Promise(function(aResolve, aReject) {
    info('Sender: --- testIncomingMessage ---');
    connection.addEventListener('message', function messageHandler(evt) {
      connection.removeEventListener('message', messageHandler);
      let msg = evt.data;
      is(msg, "msg-receiver-to-sender", "Sender: Sender should receive message from Receiver");
      postMessageToIframe('message-from-receiver-received');
      aResolve();
    });
    postMessageToIframe('trigger-message-from-receiver');
  });
}

function testTerminateConnection() {
  return new Promise(function(aResolve, aReject) {
    info('Sender: --- testTerminateConnection ---');
    connection.onstatechange = function() {
      connection.onstatechange = null;
      is(connection.state, "terminated", "Sender: Connection should be terminated.");
      aResolve();
    };
    connection.terminate();
  });
}

function teardown() {
  gScript.addMessageListener('teardown-complete', function teardownCompleteHandler() {
    gScript.removeMessageListener('teardown-complete', teardownCompleteHandler);
    gScript.destroy();
    SimpleTest.finish();
  });

  gScript.sendAsyncMessage('teardown');
}

function runTests() {
  setup()
  .then(testCreateRequest)
  .then(testStartConnection)
  .then(testSendMessage)
  .then(testIncomingMessage)
  .then(testTerminateConnection);
}

SimpleTest.waitForExplicitFinish();
SpecialPowers.pushPermissions([
  {type: 'presentation-device-manage', allow: false, context: document},
  {type: 'presentation', allow: true, context: document},
  {type: "browser", allow: true, context: document},
], () => {
  SpecialPowers.pushPrefEnv({ 'set': [["dom.presentation.enabled", true],
                                      ["dom.presentation.test.enabled", true],
                                      ["dom.presentation.test.stage", 0],
                                      ["dom.mozBrowserFramesEnabled", true],
                                      ["dom.ipc.browser_frames.oop_by_default", true]]},
    	                      runTests);
});

</script>
  </body>
</html>
